+++ crates/miden-protocol/src/batch/proposed_batch.rs
@@ -431,9 +431,10 @@ mod tests {
 
     use super::*;
     use crate::Word;
+    use crate::account::delta::AccountUpdateDetails;
     use crate::account::{AccountIdVersion, AccountStorageMode, AccountType};
     use crate::asset::FungibleAsset;
-    use crate::transaction::ProvenTransactionBuilder;
+    use crate::transaction::ProvenTransaction;
 
     #[test]
     fn proposed_batch_serialization() -> anyhow::Result<()> {
@@ -474,18 +475,20 @@ mod tests {
         let expiration_block_num = reference_block_header.block_num() + 1;
         let proof = ExecutionProof::new_dummy();
 
-        let tx = ProvenTransactionBuilder::new(
+        let tx = ProvenTransaction::new(
             account_id,
             initial_account_commitment,
             final_account_commitment,
             account_delta_commitment,
+            AccountUpdateDetails::Private,
+            vec![],
+            vec![],
             block_num,
             block_ref,
             FungibleAsset::mock(100).unwrap_fungible(),
             expiration_block_num,
             proof,
         )
-        .build()
         .context("failed to build proven transaction")?;
 
         let batch = ProposedBatch::new(