+++ crates/miden-protocol/src/transaction/mod.rs
@@ -21,12 +21,7 @@ pub use kernel::{TransactionAdviceInputs, TransactionEventId, TransactionKernel,
 pub use ordered_transactions::OrderedTransactionHeaders;
 pub use outputs::{OutputNote, OutputNotes, TransactionOutputs};
 pub use partial_blockchain::PartialBlockchain;
-pub use proven_tx::{
-    InputNoteCommitment,
-    ProvenTransaction,
-    ProvenTransactionBuilder,
-    TxAccountUpdate,
-};
+pub use proven_tx::{InputNoteCommitment, ProvenTransaction, TxAccountUpdate};
 pub use transaction_id::TransactionId;
 pub use tx_args::{TransactionArgs, TransactionScript};
 pub use tx_header::TransactionHeader;